perf: deltas Socket.IO + découpage fileParser (phase 2)
- remoteScoreServer: arena:update émis vers la room de l'arène uniquement (plus de broadcast global) ; photos base64 envoyées une seule fois par match, updates suivants allégés (~1 Ko vs ~1 Mo) ; purge du suivi photos en fin de session
- fileParser.ts (1325 l.) découpé en modules par format (txt, FFE, XML, classement, Engarde) avec barrel de ré-export — API publique inchangée, 34 tests identiques
